refactor(Problem4): Word 클래스 분리
