+++ src/main/java/onboarding/problem4/Alphabet.java
@@ -1,0 +1,33 @@
+package onboarding.problem4;
+
+public class Alphabet {
+    static final char A_LOWERCASE = 'a';
+    static final char A_UPPERCASE = 'A';
+    static final char Z_LOWERCASE = 'z';
+    static final char Z_UPPERCASE = 'Z';
+
+    public static char changeAlphabet(char alphabet) {
+        if (!Character.isAlphabetic(alphabet)) {
+            return alphabet;
+        }
+        if (isCapitalAlphabet(alphabet)) {
+            return capAlphabetReverse(alphabet);
+        }
+        return smlAlphabetReverse(alphabet);
+    }
+
+    public static boolean isCapitalAlphabet(char character) {
+        if (A_UPPERCASE <= character && character <= Z_UPPERCASE) {
+            return true;
+        }
+        return false;
+    }
+
+    public static char capAlphabetReverse(char alphabet) {
+        return (char)(Z_UPPERCASE - alphabet + A_UPPERCASE);
+    }
+
+    public static char smlAlphabetReverse(char alphabet) {
+        return (char)(Z_LOWERCASE - alphabet + A_LOWERCASE);
+    }
+}
