+++ src/main/java/onboarding/problem4/Controller.java
@@ -3,7 +3,7 @@ package onboarding.problem4;
 public class Controller {
     public static String getSolution(String word) {
         if (Validation.check(word)) {
-            String result = Word.changeWord(word);
+            String result = Word.getReverseWord(word);
             return result;
         }
         return Constant.EXCEPTION;
